pyproject.toml
[project]
name = "psynlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "beautifulsoup4==4.14.2",
    "blis==1.3.0",
    "catalogue==2.0.10",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "cloudpathlib==0.23.0",
    "confection==0.1.5",
    "cymem==2.0.11",
    "diskcache==5.6.3",
    "en-core-web-sm",
    "filelock==3.20.0",
    "fsspec==2025.9.0",
    "hf-xet==1.1.10",
    "huggingface-hub==0.35.3",
    "idna==3.11",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "langcodes==3.5.0",
    "language-data==1.3.0",
    "llama-cpp-python==0.3.2",
    "lxml==6.0.2",
    "marisa-trie==1.3.1",
    "markdown==3.9",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "murmurhash==1.0.13",
    "networkx==3.5",
    "nltk==3.9.2",
    "numpy==2.3.4",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==11.3.0",
    "preshed==3.0.10",
    "pydantic==2.12.2",
    "pydantic-core==2.41.4",
    "pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "regex==2025.9.18",
    "requests==2.32.5",
    "rich==14.2.0",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "sentence-transformers==5.1.1",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smart-open==7.3.1",
    "soupsieve==2.8",
    "spacy==3.8.7",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.5.1",
    "sympy==1.14.0",
    "thinc==8.3.6",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.1",
    "torch==2.8.0",
    "tqdm==4.67.1",
    "transformers==4.57.1",
    "typer==0.19.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wasabi==1.1.3",
    "weasel==0.4.1",
    "wrapt==1.17.3",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
